Renamed data file for product-list tests
