+++ plugins/woocommerce/tests/e2e/api-core-tests/tests/products/product-list.test.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
- const { createSampleData, deleteSampleData } = require( '../../data/products' );
+ const { createSampleData, deleteSampleData } = require( '../../data/product-list' );
  const { productsApi } = require('../../endpoints/products');
  
  /**
